// dashboard/src/app/core/interfaces/stat/graphInterface.ts
export interface GraphNamesInterface {
  carpoolers: any;
  carpoolersPerVehicule: any;
  co2: any;
  distance: any;
  operators?: any;
  petrol: any;
  trips: any;
}

export interface ChartNamesInterface {
  tripsPerDay: any;
  tripsPerMonth: any;
  tripsPerDayCumulated: any;
  tripsSubsidizedPerDay: any;
  tripsSubsidizedPerMonth: any;
  tripsSubsidizedPerDayCumulated: any;
  distancePerDay: any;
  distancePerMonth: any;
  distancePerDayCumulated: any;
  carpoolersPerMonth: any;
  carpoolersPerDayCumulated: any;
  petrolPerMonth: any;
  petrolPerDayCumulated: any;
  co2PerMonth: any;
  co2PerDayCumulated: any;
  carpoolersPerVehiculePerDayCumulated: any;
}